[RENAME] Rename Elasticsearch to Renameme
This commit is a dryrun for renaming the codebase from Elasticsearch to
the placeholder name: Renameme. Every instance of Elasticesarch, with
the exception of license headers, bugUrls, and documentation links has
been refactored to the renameme or Renameme moniker.

NOTE: this monolithic effort did not retain commit history for some
files; thus when it comes time to rename this branch should be used
for reference only.