[Version] Rebase versioning to 1.0.0 with bwc support
This commit rebases the system versioning from 7.10.3 to 1.0.0.

To do so, a new LegacyESVersion class is added to provide backward compatibility
for predecessor versions going back to 6.0.0.alpha1. The Version class was left
to rebase at 1.0.0 and Version.CURRENT is updated to the latest 1.0.0 version.

The version id is modified to include an OpenSearch id bit (bit 28) that
differentiates OpenSearch versions from Legacy versions. This is designed to
retain the semantics while complying with the version logic contract.

Signed-off-by: Nicholas Knize <nknize@amazon.com>